fix tests/examples/readme
